move mixin class to own file
